+++ packages/eyes-universal/src/debug/tracker.ts
@@ -1,9 +1,9 @@
 import type * as types from '@applitools/types'
+import type {Socket} from '../socket'
 
 type GroupHistory = {
   manager: types.Ref
-  config: Record<string, any>
-  eyes: types.Ref[]
+  eyes: (types.Ref | Error)[]
 }
 
 type TestHistory = {
@@ -14,86 +14,80 @@ type TestHistory = {
   aborted: boolean
 }
 
-export interface Tracker {
-  makeManager(config: any, manager: types.Ref): void
-  openEyes(input: any, eyes: types.Ref): void
-  check(input: any, result: any): void
-  locate(input: any, result: any): void
-  extractTextRegions(input: any, result: any): void
-  extractText(input: any, result: any): void
-  close(input: any, result: any): void
-  abort(input: any, result: any): void
-  toJSON(): Record<string, any>
-}
+export function withTracker<TSocket extends Socket>(options: {
+  socket: TSocket
+  debug?: boolean
+}): TSocket & {getHistory?(): Record<string, any>} {
+  const socket = options.socket as TSocket & {getHistory?(): Record<string, any>}
+  if (!options.debug) return socket
 
-export function makeTracker({track = false} = {}): Tracker {
   const history = {
     managers: new Map<string, GroupHistory>(),
     eyes: new Map<string, TestHistory>(),
     startedAt: new Date().toISOString(),
   }
 
+  const originalCommand = socket.command
+  socket.command = function command(name, fn) {
+    return originalCommand(name, async payload => {
+      let result, error
+      try {
+        result = await fn(payload)
+        return result
+      } catch (err) {
+        error = err
+        throw error
+      } finally {
+        if (name === 'Core.makeManager') {
+          const managerRef = result
+          history.managers.set(extractRefId(managerRef), {...payload, manager: result, eyes: []})
+        } else if (name === 'EyesManager.openEyes') {
+          const managerRef = payload.manager
+          const managerHistory = history.managers.get(extractRefId(managerRef))
+          if (error) {
+            managerHistory.eyes.push({...payload, error})
+          } else {
+            const eyesRef = result
+            managerHistory.eyes.push(eyesRef)
+            history.eyes.set(extractRefId(eyesRef), {...payload, eyes: eyesRef, commands: []})
+          }
+        } else if (/Eyes\.(check|locate|extractText|extractTextRegions)/.test(name)) {
+          const eyesRef = payload.eyes
+          const eyesHistory = history.eyes.get(extractRefId(eyesRef))
+          if (eyesHistory) {
+            const command = {name, ...payload}
+            if (error) command.error = error
+            else command.result = result
+            eyesHistory.commands.push(command)
+          }
+        } else if (/Eyes\.(close|abort)/.test(name)) {
+          const eyesRef = payload.eyes
+          const eyesHistory = history.eyes.get(extractRefId(eyesRef))
+          if (eyesHistory) {
+            eyesHistory.aborted = name.endsWith('abort')
+            eyesHistory.result = error ?? result
+          }
+        }
+      }
+    })
+  }
+
+  socket.getHistory = function getHistory() {
+    return {
+      managers: Array.from(history.managers.values(), managerMeta => ({
+        ...managerMeta,
+        eyes: managerMeta.eyes.map(eyesRefOrError => {
+          return eyesRefOrError instanceof Error ? eyesRefOrError : history.eyes.get(extractRefId(eyesRefOrError))
+        }),
+      })),
+      startedAt: history.startedAt,
+      requestedAt: new Date().toISOString(),
+    }
+  }
+
+  return socket
+
   function extractRefId(ref: types.Ref): string {
     return ref['applitools-ref-id']
   }
-
-  const api = {
-    makeManager(config: any, manager: types.Ref) {
-      history.managers.set(extractRefId(manager), {manager, config, eyes: []})
-    },
-    openEyes({manager, ...input}: any, eyes: types.Ref) {
-      const managerHistory = history.managers.get(extractRefId(manager))
-      if (!managerHistory) return
-      managerHistory.eyes.push(eyes)
-      history.eyes.set(extractRefId(eyes), {eyes, ...input, commands: []})
-    },
-    check({eyes, ...input}: any, result: any) {
-      const eyesHistory = history.eyes.get(extractRefId(eyes))
-      if (!eyesHistory) return
-      eyesHistory.commands.push({name: 'check', ...input, result})
-    },
-    locate({eyes, ...input}: any, result: any) {
-      const eyesHistory = history.eyes.get(extractRefId(eyes))
-      if (!eyesHistory) return
-      eyesHistory.commands.push({name: 'locate', ...input, result})
-    },
-    extractTextRegions({eyes, ...input}: any, result: any) {
-      const eyesHistory = history.eyes.get(extractRefId(eyes))
-      if (!eyesHistory) return
-      eyesHistory.commands.push({name: 'extractTextRegions', ...input, result})
-    },
-    extractText({eyes, ...input}: any, result: any) {
-      const eyesHistory = history.eyes.get(extractRefId(eyes))
-      if (!eyesHistory) return
-      eyesHistory.commands.push({name: 'extractText', ...input, result})
-    },
-    close({eyes}: any, result: any) {
-      const eyesHistory = history.eyes.get(extractRefId(eyes))
-      if (!eyesHistory) return
-      eyesHistory.result = result
-      eyesHistory.aborted = false
-    },
-    abort({eyes}: any, result: any) {
-      const eyesHistory = history.eyes.get(extractRefId(eyes))
-      if (!eyesHistory) return
-      eyesHistory.result = result
-      eyesHistory.aborted = true
-    },
-    toJSON() {
-      return {
-        managers: Array.from(history.managers.values(), manager => ({
-          ...manager,
-          eyes: manager.eyes.map(eyes => history.eyes.get(extractRefId(eyes))),
-        })),
-        startedAt: history.startedAt,
-        requestedAt: new Date().toISOString(),
-      }
-    },
-  }
-
-  if (!track) {
-    return Object.keys(api).reduce((api, method) => Object.assign(api, {[method]: () => void 0}), {} as Tracker)
-  }
-
-  return api
 }