+++ packages/eyes-universal/src/universal-server.ts
@@ -6,7 +6,7 @@ import {makeSpec, webdriverSpec} from './spec-driver'
 import {makeHandler} from './handler'
 import {makeSocket} from './socket'
 import {makeRefer} from './refer'
-import {makeTracker} from './debug/tracker'
+import {withTracker} from './debug/tracker'
 
 const IDLE_TIMEOUT = 900000 // 15min
 
@@ -20,9 +20,11 @@ export async function makeServer({debug = false, idleTimeout = IDLE_TIMEOUT, ...
   let idle = setTimeout(() => server.close(), idleTimeout)
 
   server.on('connection', client => {
-    const tracker = makeTracker({track: debug})
     const refer = makeRefer()
-    const socket = makeSocket(client) as Socket & types.ServerSocket<Driver, Context, Element, Selector>
+    const socket = withTracker({
+      debug,
+      socket: makeSocket(client) as Socket & types.ServerSocket<Driver, Context, Element, Selector>,
+    })
 
     clearTimeout(idle)
     socket.on('close', () => {
@@ -30,46 +32,40 @@ export async function makeServer({debug = false, idleTimeout = IDLE_TIMEOUT, ...
       idle = setTimeout(() => server.close(), idleTimeout)
     })
 
-    const init = new Promise<types.Core<Driver, Element, Selector>>(resolve => {
-      socket.once('Session.init', ({name, version, commands, protocol}) => {
-        const sdk = makeSDK<Driver, Context, Element, Selector>({
-          name: `eyes-universal/${name}`,
-          version: `${require('../package.json').version}/${version}`,
-          spec: protocol === 'webdriver' ? webdriverSpec : makeSpec({socket, commands}),
-          VisualGridClient: require('@applitools/visual-grid-client'),
-        })
-        resolve(sdk)
+    const sdkPromise = socket.create('Core.makeSDK', ({name, version, commands, protocol}) => {
+      return makeSDK<Driver, Context, Element, Selector>({
+        name: `eyes-universal/${name}`,
+        version: `${require('../package.json').version}/${version}`,
+        spec: protocol === 'webdriver' ? webdriverSpec : makeSpec({socket, commands}),
+        VisualGridClient: require('@applitools/visual-grid-client'),
       })
     })
 
     socket.command('Core.makeManager', async config => {
-      const sdk = await init
-      const manager = await sdk.makeManager(config)
-      const managerRef = refer.ref(manager)
-      tracker.makeManager(config, managerRef)
+      const sdk = await sdkPromise
+      const managerRef = refer.ref(await sdk.makeManager(config))
       return managerRef
     })
     socket.command('Core.getViewportSize', async ({driver}) => {
-      const sdk = await init
+      const sdk = await sdkPromise
       return sdk.getViewportSize({driver})
     })
     socket.command('Core.setViewportSize', async ({driver, size}) => {
-      const sdk = await init
+      const sdk = await sdkPromise
       return sdk.setViewportSize({driver, size})
     })
     socket.command('Core.closeBatches', async settings => {
-      const sdk = await init
+      const sdk = await sdkPromise
       return sdk.closeBatches(settings)
     })
     socket.command('Core.deleteTest', async settings => {
-      const sdk = await init
+      const sdk = await sdkPromise
       return sdk.deleteTest(settings)
     })
 
     socket.command('EyesManager.openEyes', async ({manager, driver, config}) => {
       const eyes = await refer.deref(manager).openEyes({driver, config})
       const eyesRef = refer.ref(eyes, manager)
-      tracker.openEyes({manager, driver, config}, eyesRef)
       return eyesRef
     })
     socket.command('EyesManager.closeAllEyes', async ({manager, throwErr}) => {
@@ -77,35 +73,25 @@ export async function makeServer({debug = false, idleTimeout = IDLE_TIMEOUT, ...
     })
 
     socket.command('Eyes.check', async ({eyes, settings, config}) => {
-      const result = await refer.deref(eyes).check({settings, config})
-      tracker.check({eyes, settings, config}, result)
-      return result
+      return refer.deref(eyes).check({settings, config})
     })
     socket.command('Eyes.locate', async ({eyes, settings, config}) => {
-      const result = refer.deref(eyes).locate({settings, config})
-      tracker.locate({eyes, settings, config}, result)
-      return result
+      return refer.deref(eyes).locate({settings, config})
     })
     socket.command('Eyes.extractTextRegions', async ({eyes, settings, config}) => {
-      const result = await refer.deref(eyes).extractTextRegions({settings, config})
-      tracker.extractTextRegions({eyes, settings, config}, result)
-      return result
+      return refer.deref(eyes).extractTextRegions({settings, config})
     })
     socket.command('Eyes.extractText', async ({eyes, regions, config}) => {
-      const result = await refer.deref(eyes).extractText({regions, config})
-      tracker.extractText({eyes, regions, config}, result)
-      return result
+      return refer.deref(eyes).extractText({regions, config})
     })
     socket.command('Eyes.close', async ({eyes, throwErr}) => {
       const results = await refer.deref(eyes).close({throwErr})
       refer.destroy(eyes)
-      tracker.close({eyes, throwErr}, results)
       return results
     })
     socket.command('Eyes.abort', async ({eyes}) => {
       const results = await refer.deref(eyes).abort()
       refer.destroy(eyes)
-      tracker.abort({eyes}, results)
       return results
     })
 
@@ -114,7 +100,7 @@ export async function makeServer({debug = false, idleTimeout = IDLE_TIMEOUT, ...
     })
 
     socket.command('Debug.getHistory', async () => {
-      return tracker.toJSON()
+      return socket.getHistory()
     })
   })