+++ packages/eyes-universal/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+- replace `Session.init` with `Core.makeSDK`
+- improve command tracking in debug mode
+- update spec drivers to match latest requirements
+- updated to @applitools/eyes-sdk-core@12.23.5 (from 12.23.1)
+- updated to @applitools/utils@1.2.3 (from 1.2.2)
+- updated to @applitools/visual-grid-client@15.8.25 (from 15.8.22)
+
 
 ## 0.1.3 - 2021/9/1